evm input module
